fix(smart-accounts): export config for external use
