+++ .changeset/yellow-snails-argue.md
@@ -1,0 +1,5 @@
+---
+'@protocolink/smart-accounts': patch
+---
+
+export config for external use
